Refactor AutoPlaylist page
